+++ demo/engine-data.js
@@ -2374,7 +2374,7 @@
    "org.kw"
   ]
  },
- "version": "2.3.0",
+ "version": "2.4.0",
  "visual_sequences": [
   [
    "rn",